refactor XA datasource package.
