+++ sharding-transaction/sharding-transaction-2pc/sharding-transaction-2pc-xa/src/main/java/io/shardingsphere/transaction/xa/jta/datasource/ShardingXADataSourceUtil.java
@@ -15,14 +15,12 @@
  * </p>
  */
 
-package io.shardingsphere.transaction.xa.jta;
+package io.shardingsphere.transaction.xa.jta.datasource;
 
 import com.atomikos.beans.PropertyException;
 import com.atomikos.beans.PropertyUtils;
 import io.shardingsphere.core.constant.DatabaseType;
 import io.shardingsphere.core.rule.DataSourceParameter;
-import io.shardingsphere.transaction.xa.convert.datasource.XADataSourceFactory;
-import io.shardingsphere.transaction.xa.convert.datasource.XAPropertiesFactory;
 
 import javax.sql.XADataSource;
 import java.util.Properties;
