+++ sharding-transaction/sharding-transaction-2pc/sharding-transaction-2pc-xa/src/main/java/io/shardingsphere/transaction/xa/jta/datasource/XAPropertiesFactory.java
@@ -15,14 +15,14 @@
  * </p>
  */
 
-package io.shardingsphere.transaction.xa.convert.datasource;
+package io.shardingsphere.transaction.xa.jta.datasource;
 
 import io.shardingsphere.core.constant.DatabaseType;
-import io.shardingsphere.transaction.xa.convert.datasource.dialect.H2XAProperties;
-import io.shardingsphere.transaction.xa.convert.datasource.dialect.MySQLXAProperties;
-import io.shardingsphere.transaction.xa.convert.datasource.dialect.OracleXAProperties;
-import io.shardingsphere.transaction.xa.convert.datasource.dialect.PostgreSQLXAProperties;
-import io.shardingsphere.transaction.xa.convert.datasource.dialect.SQLServerXAProperties;
+import io.shardingsphere.transaction.xa.jta.datasource.dialect.H2XAProperties;
+import io.shardingsphere.transaction.xa.jta.datasource.dialect.MySQLXAProperties;
+import io.shardingsphere.transaction.xa.jta.datasource.dialect.OracleXAProperties;
+import io.shardingsphere.transaction.xa.jta.datasource.dialect.PostgreSQLXAProperties;
+import io.shardingsphere.transaction.xa.jta.datasource.dialect.SQLServerXAProperties;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
