+++ sharding-transaction/sharding-transaction-2pc/sharding-transaction-2pc-xa/src/test/java/io/shardingsphere/transaction/xa/convert/datasource/XADataSourceFactoryTest.java
@@ -19,6 +19,7 @@ package io.shardingsphere.transaction.xa.convert.datasource;
 
 import com.microsoft.sqlserver.jdbc.SQLServerXADataSource;
 import io.shardingsphere.core.constant.DatabaseType;
+import io.shardingsphere.transaction.xa.jta.datasource.XADataSourceFactory;
 import org.h2.jdbcx.JdbcDataSource;
 import org.junit.Test;
 import org.postgresql.xa.PGXADataSource;
