+++ sharding-transaction/sharding-transaction-2pc/sharding-transaction-2pc-xa/src/test/java/io/shardingsphere/transaction/xa/convert/datasource/dialect/SQLServerXAPropertiesTest.java
@@ -17,6 +17,7 @@
 
 package io.shardingsphere.transaction.xa.convert.datasource.dialect;
 
+import io.shardingsphere.transaction.xa.jta.datasource.dialect.SQLServerXAProperties;
 import org.junit.Test;
 
 import java.util.Properties;
